add lib dir
